Conflicts:
	Alignment/src/alignmentUtils/SequencePair.java

+++ Alignment/src/alignmentUtils/GapFunction.java
@@ -1,15 +1,15 @@
 package alignmentUtils;
 
 public class GapFunction {
-	private int gapOpen;
-	private int gapExtend;
+	private float gapOpen;
+	private float gapExtend;
 
-	public GapFunction(int gapOpen, int gapExtend) {
+	public GapFunction(float gapOpen, float gapExtend) {
 		this.gapOpen = gapOpen;
 		this.gapExtend = gapExtend;
 	}
 
-	public int calcPenalty(boolean extend) {
+	public float calcPenalty(boolean extend) {
 		if(extend) {
 			return gapExtend;
 		}
